+++ client/src/app/apps/airlake/dataset/page.tsx
@@ -38,7 +38,7 @@ export default function Page() {
 
   const datasetTagsToDisplay = dataset?.data?.description?.split(" ").slice(0, 30).map((item: string) => {
     if (item.length > 4) {
-      return <Badge pill bg="dark" className="mt-2 me-2 top-0 end-0 ps-3 pe-3 p-2" key={Math.random().toString()}>{item}</Badge>
+      return <Badge bg="dark" className="mt-2 me-2 top-0 end-0 ps-3 pe-3 p-2" key={Math.random().toString()}>{item}</Badge>
     }
   })