+++ client/src/app/apps/swapstream/token/page.tsx
@@ -184,8 +184,8 @@ export default function page() {
           <p className="display-4">{balance} {selectedToken?.tokenSymbol}</p>
           <p className="muted-text mt-2">{selectedToken?.tokenName} Balance</p>
           <div className="mb-2">
-            <Badge bg="dark" pill className="mt-2 me-2 top-0 end-0 ps-3 pe-3 p-2">{selectedToken?.tokenSymbol}</Badge>
-            <Badge bg="dark" pill className="mt-2 me-2 top-0 end-0 ps-3 pe-3 p-2">{selectedToken?.tokensPerMatic} Tokens/MATIC</Badge>
+            <Badge bg="dark" className="mt-2 me-2 top-0 end-0 ps-3 pe-3 p-2">{selectedToken?.tokenSymbol}</Badge>
+            <Badge bg="dark" className="mt-2 me-2 top-0 end-0 ps-3 pe-3 p-2">{selectedToken?.tokensPerMatic} Tokens/MATIC</Badge>
           </div>
           <Button className="mt-2" disabled={isTxProcessing || !userState.apiKey} onClick={buyToken}>
             <Show when={!isTxProcessing}>Buy Token <ArrowRightIcon className="icon-right" /></Show>