+++ src/utils/sendEmail.ts
@@ -24,9 +24,9 @@ async function sendEmail(email: string, passKey: string) {
     const subject = "Lenstack Identity Passkey"
     const content = `
     <div style="font-family: Arial, sans-serif; color: #333; padding: 20px;">
-      <div style="background-color: #f4f4f4; padding: 20px; border-radius: 5px; text-align: center;">
+      <div style="background-color: #f4f4f4; padding: 20px; border-radius: 0.8rem; text-align: center;">
         <p style="font-size: 16px; color: #555; margin-top: 20px;">Use the below key as your Lenstack Identity Passkey. Do not share.</p>
-        <div style="display: inline-block; background-color: #e0e0e0; padding: 10px 20px; border-radius: 5px; font-size: 18px;">
+        <div style="display: inline-block; background-color: #e0e0e0; padding: 10px 20px; border-radius: 0.8rem; font-size: 18px;">
           <span style="color: #333; font-weight: bold;">${passKey}</span>
         </div>
       </div>