+++ client/src/app/dashboard/page.tsx
@@ -2,17 +2,27 @@
 import useFetch from "@/hooks/useFetch"
 import endPoints from "@/constants/apiEndpoints"
 import HTTPMethods from "@/constants/httpMethods"
-import { Fragment } from "react"
+import { Fragment, useCallback, useContext, useEffect, useState } from "react"
 import Show from "@/components/Show"
 import Loading from "@/components/Loading"
 import { Container, Row } from "react-bootstrap"
 import { GenericAppCardInterface } from "@/types/Types"
 import GenericAppCard from "@/components/GenericAppCard"
+import { AppContext } from "@/context/appStateProvider"
 
 export default function Page() {
+  const [{ globalSearchString }] = useContext(AppContext)
   const apps = useFetch("get-apps", endPoints.getPlatformConfigEndpoint, HTTPMethods.POST)
+  const [filteredApps, setFilteredApps] = useState([])
 
-  const appsToDisplay = apps?.data?.map((app: any) => {
+  useEffect(() => {
+    const filteredApps = apps?.data?.filter((app: any) =>
+      app.appName.toLowerCase().includes(globalSearchString.toLowerCase())
+    )
+    setFilteredApps(filteredApps)
+  }, [globalSearchString, apps.data])
+
+  const appsToDisplay = filteredApps?.map((app: any) => {
     const genericAppCardProps: GenericAppCardInterface = {
       badgeText: app.appStatus,
       className: app.appCategory,